requirements_simple.txt
# Simplified dependencies for FinSolve Internal Chatbot

# Core backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Frontend
streamlit==1.28.2
requests==2.31.0

# Data processing
pandas==2.1.4

# Vector database (simpler version)
chromadb==0.4.18

# Optional: LangChain (install separately if needed)
# langchain
# langchain-community
# langchain-openai